Got a program loaded and parsing instructions

+++ src/emu/instruction/access/sta.rs
@@ -1,5 +1,6 @@
 use crate::emu::{State, instruction::Operation};
 
+#[derive(Debug)]
 pub enum StoreAccumulator {
     ZeroPage { operand: u8 },
     ZeroPageX { operand: u8 },
@@ -15,7 +16,7 @@ impl Operation for StoreAccumulator {
         state
     }
 
-    fn get_size(&self) -> u8 {
+    fn get_size(&self) -> u16 {
         match self {
             Self::ZeroPage { operand: _ } => 2,
             Self::ZeroPageX { operand: _ } => 2,
