+++ src/emu/instruction/bitwise/ora.rs
@@ -1,5 +1,6 @@
 use crate::emu::{State, instruction::Operation};
 
+#[derive(Debug)]
 pub enum BitwiseOr {
     Immediate { operand: u8 },
     ZeroPage { operand: u8 },
@@ -16,7 +17,7 @@ impl Operation for BitwiseOr {
         state
     }
 
-    fn get_size(&self) -> u8 {
+    fn get_size(&self) -> u16 {
         match self {
             Self::Immediate { operand: _ } => 2,
             Self::ZeroPage { operand: _ } => 2,
