+++ src/emu/instruction/branch/beq.rs
@@ -1,5 +1,6 @@
 use crate::emu::{Operation, State};
 
+#[derive(Debug)]
 pub enum BranchIfEqual {
     Relative { operand: u8 },
 }
@@ -9,7 +10,7 @@ impl Operation for BranchIfEqual {
         state
     }
 
-    fn get_size(&self) -> u8 {
+    fn get_size(&self) -> u16 {
         match self {
             Self::Relative { operand: _ } => 2,
         }
